Refactor ENV variable access to use fetch
This commit improves the way environment variables are accessed
throughout the application by replacing instances of
`ENV["VARIABLE_NAME"] || "default_value"` with `ENV.fetch("VARIABLE_NAME", "default_value")`.
This change enhances code clarity and ensures that the specified
default value is utilized only when the environment variable is not set.

- Updated Rakefile to use fetch for database configuration
- Modified model configurations in configuration.rb to use fetch
- Refactored image_description_service.rb to utilize fetch for ollama endpoint retrieval